move qwen36 moe residency tests out of manager
